+++ src/boundless-engine/voxel/octree_node.hpp
@@ -12,29 +12,29 @@
 namespace Boundless {
     class OctreeNode {
         public:
-            OctreeNode(uint32_t locationalCode, uint32_t nodeSize);
+            OctreeNode(uint64_t locationalCode, uint64_t nodeSize);
             ~OctreeNode() {}
 
             std::size_t getDepth() const;
             glm::vec3 getChunkOffset() const;
-            std::uint32_t getSize() const;
+            std::uint64_t getSize() const;
             bool isLeaf() const;
             Voxel& getVoxelData();
-            std::uint32_t getLocationalCode() const;
+            std::uint64_t getLocationalCode() const;
             std::uint8_t getChildrenMask() const;
             void setChildrenMask(std::uint8_t childMask);
-            std::uint32_t getLOD() const;
+            std::uint64_t getLOD() const;
             std::uint8_t getFaceMask() const;
             void setFaceMask(std::uint8_t faceMask);
-            void setLOD(std::uint32_t lod);
+            void setLOD(std::uint64_t lod);
 
         private:
             Voxel m_voxel;
-            std::uint32_t m_locationalCode;
-            std::uint32_t m_nodeSize;
+            std::uint64_t m_locationalCode;
+            std::uint64_t m_nodeSize;
             std::uint8_t m_childrenMask = 0;
             std::uint8_t m_faceMask = 0;
-            std::uint32_t m_lod = 1;
+            std::uint64_t m_lod = 1u;
     };
 
 }
